Type Alias elf::endian::NativeEndian

pub type NativeEndian = LittleEndian;

A zero-sized type that always parses integers as if they’re in the compilation target’s native-endian order. This is useful for toolchain scenarios where a combiled binary knows it only wants to interpret ELF files compiled for the same target and doesn’t want the performance penalty of evaluating a match each time it parses an integer.

Aliased Type

struct NativeEndian;