Trait elf::file::Header

pub trait Header: Sized {
    type Word: ElfWord;
    fn from_slice<'a>(input: &'a [u8]) -> ElfResult<&'a Self>;
    fn parse_section<'a>(
        &'a self, 
        input: &'a [u8], 
        idx: u16
    ) -> ElfResult<&'a Section<Self::Word>>;
    fn ident(&self) -> Ident;
    fn get_type(&self) -> Type;
    fn machine(&self) -> Machine;
    fn entry_point(&self) -> usize;
    fn ph_offset(&self) -> usize;
    fn ph_count(&self) -> usize;
    fn ph_entry_size(&self) -> usize;
    fn sh_offset(&self) -> usize;
    fn sh_count(&self) -> usize;
    fn sh_entry_size(&self) -> usize;
    fn flags(&self) -> u32;
    fn sh_str_idx(&self) -> usize;

    fn sh_range(&self) -> Range<usize> { ... }
    fn ph_range(&self) -> Range<usize> { ... }
    fn section_index(&self, idx: usize) -> Range<usize> { ... }
    fn program_header_index(&self, idx: usize) -> Range<usize> { ... }
}

Trait representing an ELF File Header.

This trait allows HeaderRepr to provide a consistent API regardless of whether the header section uses 32- or 64-bit ELF words. A number of field values in the header of various sizes are converted to usize by this API so that they can be used as indices, etc.

For more information on ELF File Headers, refer to: + the ELF [specification] + the [OS Dev Wiki]

specification OS Dev Wiki

Associated Types

type Word: ElfWord

Required Methods

fn from_slice<'a>(input: &'a [u8]) -> ElfResult<&'a Self>

Attempt to extract an ELF file header from a slice of bytes.

fn parse_section<'a>(
&'a self,
input: &'a [u8],
idx: u16
) -> ElfResult<&'a Section<Self::Word>>

Attempt to extract a section header from a slice of bytes. TODO: can/should the index be usize?

fn ident(&self) -> Ident

fn get_type(&self) -> Type

fn machine(&self) -> Machine

fn entry_point(&self) -> usize

Offset of the program entry point

fn ph_offset(&self) -> usize

Offset of the start of program headers

fn ph_count(&self) -> usize

Number of program headers.

fn ph_entry_size(&self) -> usize

Size of a program header.

fn sh_offset(&self) -> usize

Offset of the start of section headers.

fn sh_count(&self) -> usize

Number of section headers.

fn sh_entry_size(&self) -> usize

Size of a section header.

fn flags(&self) -> u32

TODO: can this return the flags type?

fn sh_str_idx(&self) -> usize

Index of the section header string table.

Provided Methods

fn sh_range(&self) -> Range<usize>

fn ph_range(&self) -> Range<usize>

fn section_index(&self, idx: usize) -> Range<usize>

Calculate the index for a section header

TODO: should this check the index is reasonable & return a Result

fn program_header_index(&self, idx: usize) -> Range<usize>

Calculate the index for a program header

TODO: should this check the index is reasonable & return a Result