Trait elf::program::Header

pub trait Header: Sized {
    type Word: ElfWord;
    fn ty(&self) -> Type;
    fn offset(&self) -> usize;
    fn vaddr(&self) -> Self::Word;
    fn paddr(&self) -> Self::Word;
    fn file_size(&self) -> usize;
    fn mem_size(&self) -> usize;
    fn flags(&self) -> Flags;
    fn align(&self) -> usize;
}

Trait representing an ELF Program Header.

This trait allows HeaderRepr to provide a consistent API regardless of whether the header section uses 32- or 64-bit ELF words. A number of field values in the header of various sizes are converted to usize by this API so that they can be used as indices, etc.

For more information on ELF Program Headers, refer to: + the ELF [specification] + the [OS Dev Wiki]

specification OS Dev Wiki

Associated Types

type Word: ElfWord

Required Methods

fn ty(&self) -> Type

Returns the type of this program header.

This member tells what kind of segment this header describes or how to interpret the array element's information.

fn offset(&self) -> usize

Returns this segment's start offset from the beginning of the binary.

fn vaddr(&self) -> Self::Word

Returns the virtual address of the first byte in this segment.

fn paddr(&self) -> Self::Word

Returns the physical address of the first byte in this segment.

fn file_size(&self) -> usize

Returns the number of bytes in the file image of the segment.

This may be zero.

fn mem_size(&self) -> usize

Returns the number of bytes in the memory image of the segment.

This may be zero.

fn flags(&self) -> Flags

Returns the flags for this segment.

fn align(&self) -> usize