Trait elf::section::Header

pub trait Header: Debug {
    type Word: ElfWord;
    fn name_offset(&self) -> u32;
    fn get_type(&self) -> ElfResult<Type>;
    fn flags(&self) -> Flags;
    fn address(&self) -> PAddr;
    fn offset(&self) -> usize;
    fn length(&self) -> usize;
    fn link(&self) -> u32;
    fn info(&self) -> u32;
    fn address_align(&self) -> Self::Word;
    fn entry_length(&self) -> usize;

    fn end_address(&self) -> PAddr { ... }
    fn is_writable(&self) -> bool { ... }
    fn is_allocated(&self) -> bool { ... }
    fn is_executable(&self) -> bool { ... }
    fn is_mergeable(&self) -> bool { ... }
    fn is_uniform(&self) -> bool { ... }
    fn get_name<'a>(&self, strtab: StrTable<'a>) -> &'a str { ... }
}

Represents an ELF section header

Refer to Figure 4-8, "Section Header", from Chapter 4 of the ELF standard for more information.

Associated Types

type Word: ElfWord

Required Methods

fn name_offset(&self) -> u32

fn get_type(&self) -> ElfResult<Type>

This member categorizes the section's contents and semantics.

fn flags(&self) -> Flags

fn address(&self) -> PAddr

fn offset(&self) -> usize

fn length(&self) -> usize

TODO: should offset + length make a Range?

fn link(&self) -> u32

fn info(&self) -> u32

fn address_align(&self) -> Self::Word

fn entry_length(&self) -> usize

Provided Methods

fn end_address(&self) -> PAddr

Returns the end address of this section TODO: refactor this to return a Range instead?

fn is_writable(&self) -> bool

Returns true if this section is writable.

fn is_allocated(&self) -> bool

Returns true if this section occupies memory during program execution.

fn is_executable(&self) -> bool

Returns true if this section contains executable instructions.

fn is_mergeable(&self) -> bool

Returns true if this section can be merged.

fn is_uniform(&self) -> bool

Returns true if this section contains data that is of a uniform size.

fn get_name<'a>(&self, strtab: StrTable<'a>) -> &'a str

Look up the name of this section in the passed string table.