Struct elf::RelocKind

#[repr(transparent)]
pub struct RelocKind(_);

Implementations

impl RelocKind

RISC-V

A - the addend used to compute the value of the relocatable field

B - the base address at which the object has been loaded into memory

G - the offset into the GOT of the symbol

GOT - the address of the Global Offset Table (GOT)

L - the place of the PLT entry for a symbol

P - the place of the storage unit being relocated

S - the value of the symbol referenced by the relocation entry

Z - the size of the symbol referenced by the relocation

pub const RISCV_NONE: Self = _

pub const RISCV_32: Self = _

pub const RISCV_64: Self = _

pub const RISCV_RELATIVE: Self = _

pub const RISCV_COPY: Self = _

pub const RISCV_JUMP_SLOT: Self = _

pub const RISCV_TLS_DTPMOD32: Self = _

pub const RISCV_TLS_DTPMOD64: Self = _

pub const RISCV_TLS_DTPREL32: Self = _

pub const RISCV_TLS_DTPREL64: Self = _

pub const RISCV_TLS_TPREL32: Self = _

pub const RISCV_TLS_TPREL64: Self = _

pub const RISCV_BRANCH: Self = _

pub const RISCV_JAL: Self = _

pub const RISCV_CALL: Self = _

pub const RISCV_CALL_PLT: Self = _

pub const RISCV_GOT_HI20: Self = _

pub const RISCV_TLS_GOT_HI20: Self = _

pub const RISCV_TLS_GD_HI20: Self = _

pub const RISCV_PCREL_HI20: Self = _

pub const RISCV_PCREL_LO12_I: Self = _

pub const RISCV_PCREL_LO12_S: Self = _

pub const RISCV_HI20: Self = _

pub const RISCV_LO12_I: Self = _

pub const RISCV_LO12_S: Self = _

pub const RISCV_TPREL_HI20: Self = _

pub const RISCV_TPREL_LO12_I: Self = _

pub const RISCV_TPREL_LO12_S: Self = _

pub const RISCV_TPREL_ADD: Self = _

pub const RISCV_ADD8: Self = _

pub const RISCV_ADD16: Self = _

pub const RISCV_ADD32: Self = _

pub const RISCV_ADD64: Self = _

pub const RISCV_SUB8: Self = _

pub const RISCV_SUB16: Self = _

pub const RISCV_SUB32: Self = _

pub const RISCV_SUB64: Self = _

pub const RISCV_GNU_VTINHERIT: Self = _

pub const RISCV_GNU_VTENTRY: Self = _

pub const RISCV_ALIGN: Self = _

pub const RISCV_RVC_BRANCH: Self = _

pub const RISCV_RVC_JUMP: Self = _

pub const RISCV_RVC_LUI: Self = _

pub const RISCV_RELAX: Self = _

pub const RISCV_SUB6: Self = _

pub const RISCV_SET6: Self = _

pub const RISCV_SET8: Self = _

pub const RISCV_SET16: Self = _

pub const RISCV_SET32: Self = _

pub const RISCV_32_PCREL: Self = _

pub const RISCV_IRELATIVE: Self = _

impl RelocKind

x86_64

A - the addend used to compute the value of the relocatable field

B - the base address at which the object has been loaded into memory

G - the offset into the GOT at which the relocation entry’s symbol will reside

GOT - the address of the Global Offset Table (GOT)

L - the place of the PLT entry for a symbol

P - the place of the storage unit being relocated

S - the value of the symbol referenced by the relocation entry

Z - the size of the symbol referenced by the relocation

pub const X86_64_NONE: Self = _

pub const X86_64_64: Self = _

pub const X86_64_PC32: Self = _

pub const X86_64_GOT32: Self = _

pub const X86_64_PLT32: Self = _

pub const X86_64_COPY: Self = _

pub const X86_64_GLOB_DAT: Self = _

pub const X86_64_JUMP_SLOT: Self = _

pub const X86_64_RELATIVE: Self = _

pub const X86_64_GOTPCREL: Self = _

pub const X86_64_32: Self = _

pub const X86_64_32S: Self = _

pub const X86_64_16: Self = _

pub const X86_64_PC16: Self = _

pub const X86_64_8: Self = _

pub const X86_64_PC8: Self = _

pub const X86_64_DTPMOD64: Self = _

pub const X86_64_DTPOFF64: Self = _

pub const X86_64_TPOFF64: Self = _

pub const X86_64_TLSGD: Self = _

pub const X86_64_TLSLD: Self = _

pub const X86_64_DTPOFF32: Self = _

pub const X86_64_GOTTPOFF: Self = _

pub const X86_64_TPOFF32: Self = _

pub const X86_64_PC64: Self = _

pub const X86_64_GOTOFF64: Self = _

pub const X86_64_GOTPC32: Self = _

pub const X86_64_SIZE32: Self = _

pub const X86_64_SIZE64: Self = _

pub const X86_64_GOTPC32_TLSDESC: Self = _

pub const X86_64_TLSDESC_CALL: Self = _

pub const X86_64_TLSDESC: Self = _

pub const X86_64_IRELATIVE: Self = _

pub const X86_64_RELATIVE64: Self = _

pub const X86_64_GOTPCRELX: Self = _

pub const X86_64_REX_GOTPCRELX: Self = _