Struct elf::Sym

#[repr(C)]
pub struct Sym { /* private fields */ }

Implementations

impl Sym

pub const fn name_index(&self) -> usize

pub const fn binding(&self) -> Binding

pub const fn kind(&self) -> SymbolKind

pub const fn visibility(&self) -> Visibility

pub const fn section_index(&self) -> u16

pub const fn is_resolved(&self) -> bool

pub const fn is_absolute(&self) -> bool

pub const fn is_common(&self) -> bool

pub const fn value(&self) -> u64

Returns the symbol’s value

This value is interpreted slightly differently depending on the type of file:

Relocatable files: An offset from the begging of the symbol’s associated [section]. If the symbol refers to the common section (see [is_common()]), this value instead holds alignment constraints for the symbol.

Executable and Shared Object files: To make symbols more useful for the dynamic linker, this value contains a virtual memory address instead of a section offset. The associated section is irrelevant in this case.

pub const fn as_ptr<T>(&self) -> *const T

pub const fn as_mut_ptr<T>(&self) -> *mut T

pub const fn size(&self) -> u64

pub const fn contains_addr(&self, addr: u64) -> bool