Enum elf::endian::AnyEndian

pub enum AnyEndian {
    Little,
    Big,
}

An endian parsing type that can choose at runtime which byte order to parse integers as. This is useful for scenarios where a single compiled binary wants to dynamically interpret ELF files of any byte order.

Variants

Little

Used for a little-endian ELF structures that have been parsed with AnyEndian

Big

Used for a big-endian ELF structures that have been parsed with AnyEndian