Enum elf::DynTag

pub enum DynTag {

    Null,
    Needed,
    PltRelSize,
    PltGot,
    Hash,
    Strtab,
    Symtab,
    Rela,
    RelaSize,
    RelaEnt,
    StrtabSize,
    SymEnt,
    Init,
    Fini,
    Soname,
    Rpath,
    Symbolic,
    Rel,
    RelSize,
    RelEnt,
    PltRel,
    Debug,
    TextRel,
    JmpRel,
    BindNow,
    InitArray,
    FiniArray,
    InitArraySize,
    FiniArraySize,
    Runpath,
    Flags,
    Encoding,
    PreinitArray,
    PreinitArraySize,
    GnuHash,
    RelaCount,
    Flags1,
    EnvSpecific(isize),
    CpuSpecific(isize),
    Unknown(isize),

}

Variants

Null

Needed

PltRelSize

PltGot

Hash

Strtab

Address of the dynamic string table

Symtab

Address of the dynamic symbol table

Rela

Address of Rela type relocation table

RelaSize

Size, in bytes, of the Rela relocation table

RelaEnt

Size, in bytes, of each entry in the Rela relocation table

StrtabSize

Size, in bytes, of the dynamic string table

SymEnt

Init

Fini

Soname

Rpath

Symbolic

Rel

RelSize

RelEnt

PltRel

Debug

TextRel

JmpRel

BindNow

InitArray

Pointer to an array of pointers to initialization functions

FiniArray

Pointer to an array of pointers to termination functions

InitArraySize

Size, in bytes, of the array of initialization function pointers (InitArray)

FiniArraySize

Size, in bytes, of the array of termination function pointers (FiniArray)

Runpath

Flags

Encoding

PreinitArray

PreinitArraySize

GnuHash

RelaCount

Flags1

EnvSpecific(isize)

CpuSpecific(isize)

Unknown(isize)

Implementations

impl DynTag

pub const fn from_raw(tag: isize) -> Self