Module elf::gnu_symver

Parsing GNU extension sections for dynamic symbol versioning .gnu.version.*

Structs

SymbolDefinition

SymbolNamesIterator

SymbolRequirement

SymbolVersionTable

VerDef

The special GNU extension section .gnu.version_d has a section type of SHT_GNU_VERDEF This section shall contain symbol version definitions. The number of entries in this section shall be contained in the DT_VERDEFNUM entry of the Dynamic Section .dynamic, and also the sh_info member of the section header. The sh_link member of the section header shall point to the section that contains the strings referenced by this section.

VerDefAux

Version Definition Auxiliary Entries from the .gnu.version_d section

VerDefAuxIterator

VerDefIterator

VerNeed

The GNU extension section .gnu.version_r has a section type of SHT_GNU_VERNEED. This section contains required symbol version definitions. The number of entries in this section shall be contained in the DT_VERNEEDNUM entry of the Dynamic Section .dynamic and also the sh_info member of the section header. The sh_link member of the section header shall point to the referenced string table section.

VerNeedAux

Version Need Auxiliary Entries from the .gnu.version_r section

VerNeedAuxIterator

VerNeedIterator

VersionIndex

The special GNU extension section .gnu.version has a section type of SHT_GNU_VERSYM. This section shall have the same number of entries as the Dynamic Symbol Table in the .dynsym section. The .gnu.version section shall contain an array of elements of type Elfxx_Half (both of which are 16-bit unsigned integers).

Type Aliases

VersionIndexTable