Module elf::types

Structs

Class
Wrapper type for Class

Data
Wrapper type for Data

FileHeader
ELF File Header

Machine
Wrapper type for Machine

OSABI
Wrapper type for OSABI

ProgFlag
Wrapper type for ProgFlag

ProgType
Wrapper type for ProgType

ProgramHeader
ELF Program Header

SectionFlag
Wrapper type for SectionFlag

SectionHeader
ELF Section Header

SectionType
Wrapper type for SectionType

Type
Wrapper type for Type

Version
Wrapper type for Version This field represents the values both found in the e_ident byte array and the e_version field.

Constants

EI_ABIVERSION

EI_CLASS

EI_DATA

EI_NIDENT
Values and locations for parsing elf header ident byte array

EI_OSABI

EI_VERSION

ELFCLASS32

ELFCLASS64

ELFCLASSNONE

ELFDATA2LSB

ELFDATA2MSB

ELFDATANONE

ELFMAGIC

ELFOSABI_AIX

ELFOSABI_FREEBSD

ELFOSABI_HPUX

ELFOSABI_IRIX

ELFOSABI_LINUX

ELFOSABI_MODESTO

ELFOSABI_NETBSD

ELFOSABI_NONE

ELFOSABI_OPENBSD

ELFOSABI_SOLARIS

ELFOSABI_SYSV

ELFOSABI_TRU64

EM_386

EM_68HC05

EM_68HC08

EM_68HC11

EM_68HC12

EM_68HC16

EM_68K

EM_860

EM_88K

EM_960

EM_AARCH64

EM_ARC

EM_ARC_A5

EM_ARM

EM_AVR

EM_COLDFIRE

EM_CRIS

EM_D10V

EM_D30V

EM_FAKE_ALPHA

EM_FIREPATH

EM_FR20

EM_FR30

EM_FX66

EM_H8S

EM_H8_300

EM_H8_300H

EM_H8_500

EM_HUANY

EM_IA_64

EM_JAVELIN

EM_M32

EM_M32R

EM_ME16

EM_MICROBLAZE

EM_MIPS

EM_MIPS_RS3_LE

EM_MIPS_X

EM_MMA

EM_MMIX

EM_MN10200

EM_MN10300

EM_NCPU

EM_NDR1

EM_NONE

EM_OPENRISC

EM_PARISC

EM_PCP

EM_PDSP

EM_PJ

EM_PPC

EM_PPC64

EM_PRISM

EM_RCE

EM_RH32

EM_S370

EM_S390

EM_SH

EM_SPARC

EM_SPARC32PLUS

EM_SPARCV9

EM_ST100

EM_ST19

EM_ST7

EM_ST9PLUS

EM_STARCORE

EM_SVX

EM_TILEGX

EM_TILEPRO

EM_TINYJ

EM_TRICORE

EM_V800

EM_V850

EM_VAX

EM_VPP500

EM_X86_64

EM_XTENSA

EM_ZSP

ET_CORE

ET_DYN

ET_EXEC

ET_NONE

ET_REL

EV_CURRENT

EV_NONE

PF_NONE

PF_R

PF_W

PF_X

PT_DYNAMIC

PT_GNU_EH_FRAME

PT_GNU_RELRO

PT_GNU_STACK

PT_INTERP

PT_LOAD

PT_NOTE

PT_NULL

PT_PHDR

PT_SHLIB

PT_TLS

SHF_ALLOC

SHF_EXECINSTR

SHF_GROUP

SHF_INFO_LINK

SHF_LINK_ORDER

SHF_MERGE

SHF_NONE

SHF_OS_NONCONFORMING

SHF_STRINGS

SHF_TLS

SHF_WRITE

SHT_DYNAMIC

SHT_DYNSYM

SHT_FINI_ARRAY

SHT_GNU_ATTRIBUTES

SHT_GNU_HASH

SHT_GNU_LIBLIST

SHT_GNU_VERDEF

SHT_GNU_VERNEED

SHT_GNU_VERSYM

SHT_GROUP

SHT_HASH

SHT_INIT_ARRAY

SHT_NOBITS

SHT_NOTE

SHT_NULL

SHT_NUM

SHT_PREINIT_ARRAY

SHT_PROGBITS

SHT_REL

SHT_RELA

SHT_SHLIB

SHT_STRTAB

SHT_SYMTAB

SHT_SYMTAB_SHNDX