Struct elf::types::ProgramHeader

pub struct ProgramHeader {
    pub progtype: ProgType,
    pub offset: u64,
    pub vaddr: u64,
    pub paddr: u64,
    pub filesz: u64,
    pub memsz: u64,
    pub flags: ProgFlag,
    pub align: u64,
}

Encapsulates the contents of an ELF Program Header

The program header table is an array of program header structures describing the various segments for program execution.

Fields

progtype

Program segment type

offset

Offset into the ELF file where this segment begins

vaddr

Virtual adress where this segment should be loaded

paddr

Physical address where this segment should be loaded

filesz

Size of this segment in the file

memsz

Size of this segment in memory

flags

Flags for this segment

align

file and memory alignment